+++ tests/TestCase/TestSuite/TestCaseTest.php
@@ -1,0 +1,38 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * This file is part of me-cms.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ */
+
+namespace MeCms\Test\TestCase\TestSuite;
+
+use App\SkipTestCase;
+use MeCms\TestSuite\TestCase;
+
+/**
+ * TestCaseTest class
+ */
+class TestCaseTest extends TestCase
+{
+    /**
+     * Test for `skipIfCakeIsLessThan()` method
+     * @ŧest
+     */
+    public function testSkipIfCakeIsLessThan(): void
+    {
+        $result = (new SkipTestCase('testSkipIfCakeIsLessThanTrue'))->run();
+        $this->assertSame(1, $result->skippedCount());
+
+        $result = (new SkipTestCase('testSkipIfCakeIsLessThanFalse'))->run();
+        $this->assertSame(0, $result->skippedCount());
+    }
+}